% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dLBPG.R
\name{dLBPG}
\alias{dLBPG}
\title{The probability density function of the length-biased power Garima distribution.}
\usage{
dLBPG(x, lambda, beta)
}
\arguments{
\item{x}{vector of positive quantile.}

\item{lambda}{positive parameter(Transformed parameter).}

\item{beta}{positive parameter(Shape parameter).}
}
\value{
dLBPG gives the probability density function.
}
\description{
The LBPG package computes the probability density, the cumulative density distribution and the quantile function of the length-biased power Garima (LBPG) distribution, and generates sample values with random variables that has the LBPG distribution.
}
\examples{
dLBPG(5.7,1.5,2.5)
}
\references{
Kittipong Klinjan and Sirinapa Aryuyuen(2021), The length-biased power Garima distribution and its application to model lifetime data,Songklanakarin Lournal of Science asd Techno;ogy (SJST),Volume 43 No.3(May - Jun. 2021),pp667-676, <DOI: 10.14456/sjst-psu.2021.89>
}
