% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pLBPG.R
\name{pLBPG}
\alias{pLBPG}
\title{The cumulative density function of the length-biased power Garima distribution.}
\usage{
pLBPG(x, lambda, beta)
}
\arguments{
\item{x}{vector of positive quantile.}

\item{lambda}{positive parameters(Transformed parameter).}

\item{beta}{positive parameters(Shape parameter).}
}
\value{
pLBPG gives the cumulative density function.
}
\description{
The LBPG package computes the probability density, the cumulative density distribution and the quantile function of the length-biased power Garima (LBPG) distribution, and generates sample values with random variables that has the LBPG distribution.
}
\examples{
pLBPG(0.5,1.5,2.5)
}
\references{
Kittipong Klinjan and Sirinapa Aryuyuen(2021), The length-biased power Garima distribution and its application to model lifetime data,Songklanakarin Lournal of Science asd Techno;ogy (SJST),Volume 43 No.3(May - Jun. 2021),pp667-676, <DOI: 10.14456/sjst-psu.2021.89>
}
