% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qLBPG.R
\name{qLBPG}
\alias{qLBPG}
\title{The quantile function of the length-biased power Garima distribution.}
\usage{
qLBPG(p, alpha, beta)
}
\arguments{
\item{p}{vector pf probabilities.}

\item{alpha}{positive parameters(Transformed parameter).}

\item{beta}{positive parameters(Shape parameter).}
}
\value{
qLBPG gives the quantile function.
}
\description{
The LBPG package computes the probability density, the cumulative density distribution and the quantile function of the length-biased power Garima (LBPG) distribution, and generates sample values with random variables that has the LBPG distribution.
}
\examples{
qLBPG(0.5,1.5,2.5)
}
\references{
Kittipong Klinjan and Sirinapa Aryuyuen(2021), The length-biased power Garima distribution and its application to model lifetime data,Songklanakarin Lournal of Science asd Techno;ogy (SJST),Volume 43 No.3(May - Jun. 2021),pp667-676, <DOI: 10.14456/sjst-psu.2021.89>
}
