% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rLBPG.R
\name{rLBPG}
\alias{rLBPG}
\title{Random number generating of the length-biased power Garima distribution}
\usage{
rLBPG(n, alpha, beta)
}
\arguments{
\item{n}{number of observations.If length(n)>1,the length is taken no be the number required.}

\item{alpha}{positive parameters(Transformed parameter).}

\item{beta}{positive parameters(Shape parameter).}
}
\value{
rLBPG generates sample values of random variables.
}
\description{
The LBPG package computes the probability density, the cumulative density distribution and the quantile function of the length-biased power Garima (LBPG) distribution, and generates sample values with random variables that has the LBPG distribution.
}
\examples{
rLBPG(5,1.5,1)
}
\references{
Kittipong Klinjan and Sirinapa Aryuyuen(2021), The length-biased power Garima distribution and its application to model lifetime data,Songklanakarin Lournal of Science asd Techno;ogy (SJST),Volume 43 No.3(May - Jun. 2021),pp667-676, <DOI: 10.14456/sjst-psu.2021.89>
}
